backend/requirements.txt
addict==2.4.0
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
autodistill==0.1.26
autodistill-grounding-dino==0.1.4
billiard==4.2.0
celery==5.4.0
certifi==2023.7.22
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
contourpy==1.2.1
cycler==0.10.0
defusedxml==0.7.1
Django==5.0.6
django-cors-headers==4.3.1
djangorestframework==3.15.1
filelock==3.14.0
fonttools==4.53.0
fsspec==2024.6.0
huggingface-hub==0.23.3
idna==2.10
importlib_metadata==7.1.0
Jinja2==3.1.4
kiwisolver==1.4.5
kombu==5.3.7
MarkupSafe==2.1.5
matplotlib==3.9.0
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105
opencv-python==4.10.0.82
opencv-python-headless==4.8.0.74
packaging==24.0
pandas==2.2.2
pillow==10.3.0
platformdirs==4.2.2
prompt_toolkit==3.0.46
psutil==5.9.8
py-cpuinfo==9.0.0
pycocotools==2.0.7
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-magic==0.4.27
pytz==2024.1
PyYAML==6.0.1
redis==5.0.5
regex==2024.5.15
requests==2.32.3
requests-toolbelt==1.0.0
rf-segment-anything==1.0
rf_groundingdino==0.1.2
roboflow==1.1.31
safetensors==0.4.3
scipy==1.13.1
seaborn==0.13.2
six==1.16.0
sqlparse==0.5.0
supervision==0.21.0
sympy==1.12.1
timm==1.0.3
tokenizers==0.19.1
tomli==2.0.1
torch==2.3.1
torchvision==0.18.1
tqdm==4.66.4
transformers==4.41.2
triton==2.3.1
typing_extensions==4.12.1
tzdata==2024.1
ultralytics==8.2.28
ultralytics-thop==0.2.7
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
yapf==0.40.2
zipp==3.19.2
